+++ src/ui/text/input/mod.rs
@@ -7,17 +7,21 @@
 mod _tests;
 
 mod define; // TextInput[Config|Action|Reject|View]
-
 // impls
 mod _helper;
 mod core;
 #[cfg(feature = "alloc")]
 mod alloc;
 
+#[cfg(feature = "event")]
+mod event; // WIP TextInputKeymap[Preset]
+
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             define::*,
         };
+        #[cfg(feature = "event")]
+        pub use super::event::*;
     }
 }